+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 next version
 ------------
 
+New features
+============
+
+* Support `DataFrame` as input to `run_amd`
+  
 Changes
 =======
 
@@ -14,6 +19,7 @@ Bugfixes
 * Fix NONMEM parsing issues where having WRES in $TABLE could lead to parsing other columns incorrectly
 * Let translate_nmtran_time return the input model if the input model has no dataset
 * Fix bug causing NONMEM code to keep bounds on thetas after unconstrain_parameters
+* Fix crashes when starting amd using the command line interface
 
 1.2.0 (2024-08-22)
 ------------------